This version of MiniCache supports SHFS 1.06,
comes with a telnet shell, and an http server
that directly operates on top of SHFS. Cache
actions can be triggered via XenStore and
download statistics are collected for each
object on hits and misses.

Please consider this release as
experimental.

+++ shfs-tools/shfs_mkfs.h
@@ -21,7 +21,7 @@ struct args {
 
 	uint8_t  allocator;
 	uint8_t  hashfunc;
-	uint32_t hashlen;
+	uint8_t  hashlen;
 	uint32_t bucket_count;
 	uint32_t entries_per_bucket;
 };
